FP-1297 : Call logout method on click without event parameter
