+++ index.js
@@ -18,7 +18,7 @@ import Themes from "./src/styles/Themes";
 import Style from "./src/styles/Style";
 import Logs from "./src/Components/Logs/Logs";
 import LoginForm from "./src/Components/LoginForm/LoginForm";
-import ProfileMenu from "./src/Components/ProfileMenu";
+import ProfileMenu from "./src/Components/ProfileMenu/ProfileMenu";
 import FilterIcon from "./src/Components/Logs/LogsFilterBar/FiltersIcon/FiltersIcon";
 import SelectScopeModal from "./src/Components/Modal/SelectScopeModal";
 import HTMLPopper from "./src/Components/Popper/HTMLPopper";
