+++ stories/ProfileMenu.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import ProfileMenu from "../src/Components/ProfileMenu";
+import ProfileMenu from "../src/Components/ProfileMenu/ProfileMenu";
 import Themes from "../src/styles/Themes";
 import { ThemeProvider } from "@material-ui/styles";
